Login.Registration UI Done
